fetch2/npm.py: refactor the npm fetcher
This commit refactors the npm fetcher to improve some points and fix
others:

The big change is that the fetcher is only fetching the package source
and no more the dependencies. This is to act as the other fetchers
(git, wget, ...). The dependencies will now be fetched by the npm class.

This commit also fixes a lot of issues with the package names (exotic
characters, scoped packages) which were badly handled.

The validation files - lockdown.json and npm-shrinkwrap.json - are no
more used by the fetcher. Instead, the downloaded tarball is validated
with the shasum provided in the npm view of the package. The shrinkwrap
file is now used by the npm class.

Signed-off-by: Jean-Marie LEMETAYER <jean-marie.lemetayer@savoirfairelinux.com>